# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 🚀 Edu Management API
  description: Education Management System API collection
  version: 1.0.0
servers:
  - url: http://{{base_url}}{{api_version}}
  - url: http://{{base_url}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
...
paths:
...
  /enrollments/:
    post:
      tags:
        - 📑 Enrollment
      summary: Enroll Student
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"student_id\": \"STUDENT_ID_HERE\",\n 
                \"class_section_id\": \"CLASS_SECTION_ID_HERE\",\n 
                \"roll_number\": 1,\n  \"enrollment_date\":
                \"2024-06-01T00:00:00Z\"\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /enrollments/{id}:
    get:
      tags:
        - 📑 Enrollment
      summary: Get Enrollment By ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - 📑 Enrollment
      summary: Delete Enrollment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /enrollments/student/{student_id}:
    get:
      tags:
        - 📑 Enrollment
      summary: Get Enrollments By Student
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
        - name: limit
          in: query
          schema:
            type: integer
          example: '10'
        - name: student_id
          in: path
          schema:
            type: string
          required: true
          example: STUDENT_ID_HERE
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /enrollments/class/{class_section_id}/roster:
    get:
      tags:
        - 📑 Enrollment
      summary: Get Class Roster
      security:
        - bearerAuth: []
      parameters:
        - name: class_section_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /enrollments/{id}/status:
    patch:
      tags:
        - 📑 Enrollment
      summary: Withdraw Student
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"status\": \"WITHDRAWN\",\n  \"left_date\":
                \"2024-11-01T00:00:00Z\"\n}"
      security:
        - bearerAuth: []
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /exams/:
    post:
      tags:
        - 🎓 Exam
      summary: Create Exam
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"academic_year_id\": \"{{academic_year_id}}\",\n 
                \"name\": \"Midterm Examination 2025\",\n  \"description\":
                \"First semester midterm exams\",\n  \"exam_type\":
                \"MIDTERM\",\n  \"start_date\": \"2025-09-15T00:00:00Z\",\n 
                \"end_date\": \"2025-09-25T00:00:00Z\"\n}"
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      tags:
        - 🎓 Exam
      summary: List Exams
      parameters:
        - name: academic_year_id
          in: query
          schema:
            type: string
          example: '{{academic_year_id}}'
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
        - name: page_size
          in: query
          schema:
            type: integer
          example: '10'
        - name: sort_by
          in: query
          schema:
            type: string
          example: start_date
        - name: sort_dir
          in: query
          schema:
            type: string
          example: asc
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /exams/{exam_id}:
    get:
      tags:
        - 🎓 Exam
      summary: Get Exam by ID
      parameters:
        - name: exam_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    put:
      tags:
        - 🎓 Exam
      summary: Update Exam
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"name\": \"Midterm Examination 2025 (Updated)\",\n 
                \"end_date\": \"2025-09-27T00:00:00Z\"\n}"
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: exam_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    delete:
      tags:
        - 🎓 Exam
      summary: Delete Exam
      parameters:
        - name: exam_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /exams/{exam_id}/publish:
    patch:
      tags:
        - 🎓 Exam
      summary: Unpublish Exam
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: '"{\n  \"is_published\": false\n}"'
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: exam_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /exams/{exam_id}/schedule:
    post:
      tags:
        - 📅 Exam Schedule
      summary: Create Schedule
      requestBody:
        content:
          '*/*':
            schema:
              type: string
              example: >-
                "{\n  \"class_section_id\": \"{{class_section_id}}\",\n 
                \"subject_id\": \"{{subject_id}}\",\n  \"exam_date\":
                \"2025-09-17T00:00:00Z\",\n  \"start_time\":
                \"2025-09-17T09:00:00Z\",\n  \"end_time\":
                \"2025-09-17T11:00:00Z\",\n  \"max_marks\": \"100.00\",\n 
                \"passing_marks\": \"35.00\",\n  \"room_number\": \"Room
                101\"\n}"
      parameters:
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: exam_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    get:
      tags:
        - 📅 Exam Schedule
      summary: List Schedules by Exam
      parameters:
        - name: exam_id
          in: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...